refactor(agent): extract Anthropic usage/pricing into pure functions
Begin decomposing the modelHandlerAnthropic god-file (~2.1k LOC) by
pulling its usage-accounting and cache-aware pricing concern into a
dedicated anthropic/anthropicUsage.ts:

- getAnthropicUsageTokenTotals  (per-iteration vs flat usage totals)
- computeAnthropicPrice         (base + cache-creation/read pricing)
- normalizeAnthropicUsage       (unified NormalizedUsage shape)

These are pure functions of the raw usage plus an AnthropicPricingConfig
(inputPrice/outputPrice/supportsPromptCaching/cacheDiscountFactor), so the
pricing math is now unit-testable without a live handler. The handler keeps
its computePrice/normalizeUsage overrides as thin delegators that pass
this.config/this.capabilities — the IModelHandler surface is unchanged.

No behavior change. Verified: typecheck, lint (0 errors), and the full
test:kernel suite (1383 tests, incl. Anthropic cache-pricing) pass.
